F: add file_sd_configs for axe-rest-api
Though it is not the cleanest, it gets the jobs done. Further changes to
axe-rest-api deployment configuration should be synced to
prometheus-deployment.yaml.

Though note that this tries to scrape ip address of cluster host, so if
your axe-rest-api server is running on baremetal but your cluster is
running on vm, it would not be able to scrape metrics. Deploy cluster on
baremetal or run your server on top of vm.